requirements.txt
Django==1.10.1
djangorestframework==3.4.6
requests==2.11.1
httpretty==0.8.14
mock==2.0.0